refactor(utility): Function for PC Wifi IP identification moved to utility.py
